% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_dataset.R
\name{summarize_subject_per_category}
\alias{summarize_subject_per_category}
\title{Summarise subject level info per cateogry
'}
\usage{
summarize_subject_per_category(dt, check = TRUE, verbose = TRUE,
  time_breaks = seq(from = 1, by = 7, length.out = 2 * 52 + 1))
}
\description{
Summarise subject level info per cateogry
'
}

