% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_dataset.R
\name{summarize_dataset}
\alias{summarize_dataset}
\alias{summarize_dataset_json}
\title{Summarize full dataset}
\usage{
summarize_dataset(dt, check = TRUE, group_duration = "week",
  verbose = TRUE, agedays_min = -365, agedays_max = 265 * 2)

summarize_dataset_json(..., pretty = TRUE)
}
\arguments{
\item{dt}{dataset to summarizes}

\item{check}{boolean to determine if \code{check_data()} should be performed}

\item{group_duration}{string of one of \code{c("week", "month", "quarter", "year")}}

\item{verbose}{boolean to determine if progress bars should be displayed}

\item{agedays_min, agedays_max}{min and max agedays allowed}

\item{...}{all args directly passed to summarize_dataset}

\item{pretty}{boolean to determine if the json should be pretty printed}
}
\description{
Create a full object that summarizes both subject level and time varying variables.
}

