% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_dataset.R
\name{to_json}
\alias{to_file}
\alias{to_json}
\title{Save methods}
\usage{
to_json(x, pretty)

to_file(x, file, pretty)
}
\arguments{
\item{x}{item in question}

\item{pretty}{boolean to pretty print the json}

\item{file}{file to print the json}
}
\description{
Save methods
}

